Moved VecT hash to new noise_internals.h
